Safari-609.1.11.1/LayoutTests/requestidlecallback/requestidlecallback-document-gc-expected.txt

This tests that documents with pending requestIdleCallback is eventually collected

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS initialCount + frameCount / 2 > internals.numberOfLiveDocuments() is true
PASS successfullyParsed is true

TEST COMPLETE
